# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Qwen-2.5-VL API",
        "description": "API for Qwen-2.5-VL model with image and text inputs",
        "version": "0.0.1"
    },
    "paths": {
...
        "/patient/": {
            "get": {
                "summary": "Get Patients",
                "operationId": "get_patients_patient__get",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Patient",
                "operationId": "create_patient_patient__post",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "",
                            "title": "Patient Id"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "",
                            "title": "Name"
                        }
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "description": "",
                            "title": "Age"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "",
                            "title": "Gender"
                        }
                    },
                    {
                        "name": "dob",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "",
                            "title": "Dob"
                        }
                    },
                    {
                        "name": "medical_history",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "",
                            "title": "Medical History"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/patient/{patient_id}": {
            "get": {
                "summary": "Get Patient",
                "operationId": "get_patient_patient__patient_id__get",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Patient Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Patient",
                "operationId": "update_patient_patient__patient_id__put",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Patient Id"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Updated name of the patient",
                            "title": "Name"
                        },
                        "description": "Updated name of the patient"
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "description": "Updated age of the patient",
                            "title": "Age"
                        },
                        "description": "Updated age of the patient"
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "",
                            "title": "Gender"
                        }
                    },
                    {
                        "name": "dob",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "",
                            "title": "Dob"
                        }
                    },
                    {
                        "name": "medical_history",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "",
                            "title": "Medical History"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Patient",
                "operationId": "delete_patient_patient__patient_id__delete",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Patient Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "JWTBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}